+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/PlayerMixin.java
@@ -1,5 +1,6 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
+import insane96mcp.iguanatweaksreborn.effect.Pumped;
 import insane96mcp.iguanatweaksreborn.module.combat.feature.Shields;
 import insane96mcp.iguanatweaksreborn.module.experience.feature.PlayerExperience;
 import insane96mcp.iguanatweaksreborn.module.hungerhealth.feature.ExhaustionIncrease;
@@ -47,7 +48,9 @@ public abstract class PlayerMixin extends LivingEntity {
 
 	@ModifyVariable(method = "causeFoodExhaustion", argsOnly = true, at = @At("HEAD"))
 	private float applyHungerToFoodExhaustion(float amount) {
-		return ExhaustionIncrease.increaseHungerEffectiveness((Player) (Object) this, amount);
+		float newAmount = ExhaustionIncrease.increaseHungerEffectiveness((Player) (Object) this, amount);
+		newAmount = Pumped.decreaseExhaustionConsumption((Player) (Object) this, newAmount);
+		return newAmount;
 	}
 
 	@Inject(method = "disableShield", at = @At("HEAD"), cancellable = true)