+++ changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Upcoming
-* TODO Add an exhaustion reduction method
+* Added "Pumped" effect, unused as now
+  * Decreases hunger consumption
 * Added `Arrows don't trigger invincibility frames`.
   * Like combat test snapshots, hitting a mob with multiple arrows will no longer bounce them off but actually hit it.
 * Removed Regeneration from Sleeping Effect
@@ -29,7 +30,7 @@
 * Farming
   * Disabled hoes are no longer damaged when trying to till. They can still be used to break blocks.
   * Plants growth multipliers are now also affected by sunlight and nighttime
-    * Cactuses now grow much slower outside their biome
+    * Cacti now grow much slower outside their biome
   * Fixed harder crops applying hardness to non-insta-break blocks too
 * Hunger & Health
   * 'Food Consuming' and 'Food Hunger' have been merged to a single feature 'Food'